reduce compiling warning for MSVC.
below files add definition _CRT_SECURE_NO_WARNINGS for MSVC:
myapp-cxx.cpp
sqtoolcpp.cpp
test-container-cxx.cpp
test-cxx.cpp

+++ examples/myapp-c99.c
@@ -77,10 +77,10 @@ int  main(void)
 		strcut User defined in  sqapp/CStructs.h
 	 */
 	SqStorage *storage = myapp->base.storage;
-	User  user = {0};
-	User *user_ptr;
-	int   id[2];
-	int   n;
+	User     user = {0};
+	User    *user_ptr;
+	int64_t  id[2];
+	int      n;
 
 	user.age   = 18;
 	user.email = "alex@guest";
